Use reagent-forms for event creation form
Reagent-forms is nice because it handles binding data to an atom for us,
so we don't need to write a bunch of boilerplate. Moreover it allows us
to prevent cluttering our event bus with a bunch of tiny messages.

This might (inevitably will) lead to some synchronicity bugs, but the
idea is that we'll have this locally mutating state in a component and
then dispatch it out to the rest of the application on some event (like
clicking the submit button)